internal: move DialOptions to a new file
wrap comment lines

remove redundant defaultDialOptions call

rename file
